ai-analysis-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
wordcloud==1.9.2
matplotlib==3.8.2
numpy==1.24.3
pandas==2.1.4
jieba==0.42.1
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0